moss-map.pl - implement a 'bulk' api designed to support the existing app
i.e. Permits uploading a csv file of records defining a set, querying
existing data sets, deleting existing data sets, and getting a single
data set in the specific format used by the app currently (although
clearly it could be simplified, that's a job for later).

The 'comleted' arrayref is left empty, this is still to be
implemented (there are no tables for completed tetrads yet).